Fix for autosaving
Patch by Beelzy:

Don't know if you will encounter this problem, but on my system,
clicking "discard all" or sometimes "ok" causes Krita to crash. It
appears to be related to this problem:

https://blogs.kde.org/2009/03/26/how-crash-almost-every-qtkde-application-and-how-fix-it-0

Not entirely sure if using QPointer is an adequate fix. It will prevent
the crashes, if you do a check for the variables, but then you won't
obtain the files the user wanted to discard or keep. So I opted to use
signals and slots like we do in BlackBerry 10, and it won't crash
anymore. The only problem with this solution is that it looks ugly, and
requires taking certain variables off the heap. Not entirely sure if
that's desirable or acceptable.

The other issue is that checkboxes in the AutoSave dialog don't maintain
state because the goya:creatingWidgetForIndex property appears to have
disappeared from KF5 KItemViews. I'm unsure exactly why the index
parameter is commented out; maybe it didn't work at some point, but I
uncommented it and it appears to work.